Rename tabs components
