+++ myuw_vue/components/_common/visual_schedule/schedule.vue
@@ -20,12 +20,12 @@
            v-model="tabIndex"
            pills bottom-border nav-wrapper-class="mb-3 p-0">
         <template #tabs>
-          <uw-tab-list-button v-for="(period, i) in periods" :key="i" 
+          <uw-tab-button v-for="(period, i) in periods" :key="i" 
               :panel-id="period.id"
               title-item-class="myuw-text-xs me-2 mb-1"
               title-link-class="rounded-0 text-body">
             {{period.title}}
-          </uw-tab-list-button>
+          </uw-tab-button>
         </template>
         <template #panels>
           <uw-tab-panel v-for="(period, i) in periods" :key="i" :panel-id="period.id">
@@ -68,9 +68,9 @@
 <script>
 import {mapGetters, mapState, mapActions} from 'vuex';
 import Card from '../../_templates/card.vue';
-import Tabs from '../../_templates/tabs/tab-container.vue';
-import TabListButton from '../../_templates/tabs/tab-list-button.vue';
-import TabPanel from '../../_templates/tabs/tab-panel.vue';
+import Tabs from '../../_templates/tabs/tabs.vue';
+import TabButton from '../../_templates/tabs/button.vue';
+import TabPanel from '../../_templates/tabs/panel.vue';
 import ScheduleTab from './schedule-tab.vue';
 import CourseSection from './course-section.vue';
 
@@ -78,7 +78,7 @@ export default {
   components: {
     'uw-card': Card,
     'uw-tabs': Tabs,
-    'uw-tab-list-button': TabListButton,
+    'uw-tab-button': TabButton,
     'uw-tab-panel': TabPanel,
     'uw-schedule-tab': ScheduleTab,
     'uw-course-section': CourseSection,
