+++ myuw_vue/components/_templates/tabs/panel.vue
