+++ myuw_vue/components/_templates/tabs/tabs.vue
